+++ eslint.config.mjs
@@ -1,0 +1,24 @@
+import eslintKongUiConfig from '@kong/eslint-config-kong-ui'
+import eslintKongUiConfigCypress from '@kong/eslint-config-kong-ui/cypress'
+
+export default [
+  ...eslintKongUiConfig,
+  ...eslintKongUiConfigCypress.map(config => ({
+    ...config,
+    files: [
+      '**/cypress/**',
+      '**/*.cy.{js,ts,jsx,tsx}',
+    ],
+  })),
+  {
+    ignores: ['**/__template__/**'],
+  },
+  {
+    files: [
+      'packages/portal/document-viewer/src/components/**/*.vue',
+    ],
+    rules: {
+      'vue/multi-word-component-names': ['off'],
+    },
+  },
+]