+++ packages/analytics/metric-cards/src/constants.ts
@@ -1,8 +1,8 @@
-/* eslint-disable no-unused-vars */
+
 export enum MetricCardSize {
   Small = 'sm',
   Medium = 'md',
   Large = 'lg',
   LargeCompact = 'lgc',
-  ExtraLarge = 'xl'
+  ExtraLarge = 'xl',
 }