+++ packages/entities/entities-plugins/src/components/PluginEntityForm.vue
@@ -134,8 +134,9 @@ const props = defineProps({
 
 const { axiosInstance } = useAxios(props.config?.axiosRequestConfig)
 
-const { parseSchema } = composables.useSchemas(props.entityMap.focusedEntity?.id || undefined, {
-  groupFields: props.config.groupFields,
+const { parseSchema } = composables.useSchemas({
+  entityId: props.entityMap.focusedEntity?.id || undefined,
+  credential: props.credential,
   useRLARedesignedForm: props.config.useRLARedesignedForm,
 })
 const { convertToDotNotation, unFlattenObject, isObjectEmpty, unsetNullForeignKey } = composables.usePluginHelpers()
@@ -603,7 +604,9 @@ watch(() => props.schema, (newSchema, oldSchema) => {
 
   Object.assign(formModel, form.model)
 
-  formSchema.value = { fields: formSchema.value?.fields?.map((r: Record<string, any>) => { return { ...r, disabled: r.disabled || false } }) }
+  formSchema.value = { fields: formSchema.value?.fields?.map((r: Record<string, any>) => {
+    return { ...r, disabled: r.disabled || false }
+  }) }
   Object.assign(originalModel, JSON.parse(JSON.stringify(form.model)))
   sharedFormName.value = getSharedFormName(form.model.name, { useRLARedesignedForm: props.config.useRLARedesignedForm })