+++ packages/entities/entities-plugins/src/types/plugin-form.ts
@@ -36,8 +36,6 @@ export interface BasePluginFormConfig {
   entityId?: string
   /** Whether to hide the consumer group scope field. For Kong Manager OSS, this is true */
   disableConsumerGroupScope?: boolean
-  /** Whether to enable grouping for required and advanced (optional) fields. Default: false */
-  groupFields?: boolean
   /** Whether to use the redesigned form for the RLA plugin. Default: false */
   useRLARedesignedForm?: boolean
   /** Whether to use the horizontal radios in the plugin scoping section. Default: false */
@@ -113,7 +111,7 @@ export interface DefaultPluginsFormSchema {
   getColumnFields?: (schema: unknown) => object
 }
 
-export type PartiallyRequired<T, K extends keyof T> = { [k in K]-?: T[k] } & { [k in keyof T]: T[k] };
+export type PartiallyRequired<T, K extends keyof T> = { [k in K]-?: T[k] } & { [k in keyof T]: T[k] }
 
 export type GetRequiredFieldsByContext<T extends DefaultPluginsFormSchema> = T['type'] extends 'input' ? PartiallyRequired<DefaultPluginsFormSchema, 'inputType'> : DefaultPluginsFormSchema