+++ packages/entities/entities-plugins/src/types/plugins/acl.d.ts
@@ -7,7 +7,7 @@ interface ACLFieldSchema {
       group: {
         hint: string,
       }
-    }
+    },
   ]
 }