+++ packages/entities/entities-plugins/src/types/plugins/basic-auth.d.ts
@@ -9,7 +9,7 @@ interface BasicAuthFieldSchema {
       password: {
         inputType: string
       }
-    }
+    },
   ]
 }