+++ packages/entities/entities-plugins/src/types/plugins/hmac-auth.d.ts
@@ -13,7 +13,7 @@ interface HmacAuthFieldSchema {
         submitWhenNull?: boolean
         hint: string
       }
-    }
+    },
   ]
 }