+++ packages/entities/entities-shared/src/types/utils.ts
@@ -1,6 +1,6 @@
 import type { Ref } from 'vue'
 
-export type MaybeRef<T> = T | Ref<T>;
+export type MaybeRef<T> = T | Ref<T>
 
 export interface FilterKeys {
   fetchedItemsKey: string